requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.2
Celery==5.3.6
redis==5.0.1
requests==2.31.0
celery-redbeat==2.1.1